+++ autode/wrappers/basis_sets.py
@@ -24,6 +24,7 @@ def2ecp = ECP(name='def2-ECP',
                         'Lu: 10.1063/1.1406535'],
               orca=None,              # def2-ECP is applied by default
               nwchem='def2-ecp',
+              qchem='def2-ecp',
               min_atomic_number=37)   # applies to Rb and heavier